% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\docType{package}
\name{crane.reporter-package}
\alias{crane.reporter}
\alias{crane.reporter-package}
\title{crane.reporter: Export gtsummary Tables with reporter}
\description{
Convert gtsummary tables and plain data frames into clinical-style reporter RTF
and TXT outputs with controlled widths, pagination, indentation, spanning
headers, and optional intermediate data exports.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/openpharma/crane.reporter/}
  \item Report bugs at \url{https://github.com/openpharma/crane.reporter/issues}
}

}
\author{
\strong{Maintainer}: Joe Zhu \email{sha.joe.zhu@gmail.com}

Authors:
\itemize{
  \item Xiecheng Gu
  \item Joe Zhu \email{sha.joe.zhu@gmail.com}
  \item Bill Huang
}

}
