% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\docType{package}
\name{derrick-package}
\alias{derrick}
\alias{derrick-package}
\title{derrick: R Package Template}
\description{
R package template with GitHub Actions workflows included.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/openpharma/derrick/}
  \item Report bugs at \url{https://github.com/openpharma/derrick/issues}
}

}
\author{
\strong{Maintainer}: Xiecheng Gu

Authors:
\itemize{
  \item Joe Zhu \email{sha.joe.zhu@gmail.com}
  \item Bill Huang
}

}
