% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\docType{package}
\name{derrick-package}
\alias{derrick}
\alias{derrick-package}
\title{Export gtsummary tables with reporter}
\description{
Convert \code{gtsummary} tables and plain data frames into clinical-style
\code{reporter} RTF, TXT, DOCX, PDF, and HTML outputs.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/openpharma/derrick/}
  \item Report bugs at \url{https://github.com/openpharma/derrick/issues}
}

}
\author{
\strong{Maintainer}: Xiecheng Gu \email{guxc0405med@163.com}

Authors:
\itemize{
  \item Joe Zhu \email{sha.joe.zhu@gmail.com}
  \item Bill Huang \email{Bill.Huang@Toastmasters.org.tw}
}

}
