% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\docType{package}
\name{derrick-package}
\alias{derrick}
\alias{derrick-package}
\title{derrick: Export gtsummary Tables with reporter}
\description{
Convert gtsummary tables and plain data frames into clinical-style reporter
RTF, TXT, DOCX, PDF, and HTML outputs with controlled widths, pagination,
indentation, spanning headers, and optional intermediate data exports.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/openpharma/derrick/}
  \item Report bugs at \url{https://github.com/openpharma/derrick/issues}
}

}
\author{
\strong{Maintainer}: Joe Zhu \email{sha.joe.zhu@gmail.com}

Authors:
\itemize{
  \item Xiecheng Gu
  \item Joe Zhu \email{sha.joe.zhu@gmail.com}
  \item Bill Huang
}

}
